Readme cleanup.
There is a lack of a backtick which is fucking up the "how it works" paragraph. So I fixed that. And then a superfluous apostrophe. And then I syntax highlighted some shit because that's what you do.

### also

This project is a work of art. Thank you for this contribution to the internet.